Add link warning to bottom of page
This is to warn people that clicking on links may not produce results at the moment, but may do so in the future.  The aim is to reduce surprise and possible frustration when clicking on links to repos or pages doesn't produce relevant results.

I'd like to retrofit this to all the new format lreq docs this week, if approved, so i'd appreciate a quick review.